+++ src/pyjams/__init__.py
@@ -50,6 +50,8 @@ from . import functions
 
 # isotopic fractionation factors during liquid-water vapour equilibration
 from .alpha_equ_h2o import alpha_equ_h2o
+# kinetic fractionation of molecular diffusion of water vapour.
+from .alpha_kin_h2o import alpha_kin_h2o
 # argmax, argmin and argsort for array_like and Python iterables
 from .argsort import argmax, argmin, argsort
 # (index of) closest element in an array
@@ -67,7 +69,7 @@ from .tee import tee
 
 __all__ = ["__version__", "__author__",
            "const", "functions",
-           "alpha_equ_h2o",
+           "alpha_equ_h2o", "alpha_kin_h2o",
            "argmax", "argmin", "argsort",
            "closest",
            "division", "div",